+++ Groot/src/main/groovy/net/milosvasic/groot/Groot.groovy
@@ -2,6 +2,7 @@ package net.milosvasic.groot
 
 import groovy.util.slurpersupport.GPathResult
 import net.milosvasic.groot.deployment.Deploy
+import net.milosvasic.groot.languages.groovy.Groovy
 import net.milosvasic.groot.languages.java.Java
 import net.milosvasic.groot.languages.kotlin.Kotlin
 import org.gradle.api.Project
@@ -10,6 +11,7 @@ import org.gradle.api.Project
 class Groot {
 
     public Java java
+    public Groovy groovy
     public Kotlin kotlin
     private Project project
     public Deploy deployment
@@ -23,6 +25,7 @@ class Groot {
         this.project = project
         java  = new Java(project)
         kotlin = new Kotlin(project)
+        groovy = new Groovy(project)
         deployment = new Deploy(project)
     }